Add environment tag to Terraform workflow
Updated the GitHub Actions Terraform workflow to include environment tags. This now includes the setting of environment tag as an environment variable and adding a condition for destruction only when the environment tag is 'prod'. Furthermore, the output of the environment tag has been added to the main Terraform file.